Renamed controls-directory to PascalCase
