+++ src/lib/components/NewLayeredMap/components/Controls/components/DrawControls.js
