+++ src/lib/components/NewLayeredMap/components/Controls/components/MousePosition.js
