+++ src/lib/components/NewLayeredMap/components/Controls/components/Switch/L.SwitchControl.js
