+++ src/lib/components/NewLayeredMap/components/Controls/components/Switch/Switch.js
