+++ src/lib/components/NewLayeredMap/components/Controls/components/VerticalZoom/L.VerticalZoom.js
