+++ src/lib/components/NewLayeredMap/components/Controls/components/VerticalZoom/VerticalZoom.js
